+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [builder] Removed useless download.kiwix.org special behavior for reader checksum. Checksum now on Reader
+- [inputs.base] Version-only base image def now targets uncompressed URL
 
 ## [2.0.0] - 2024-04-16