+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.23] - 2024-04-26
+## [2.2.4] - 2024-05-01
+
+### Changed
+
+- [builder] Using captive-portal 1.4.3 fixing 2.2.1 regression
+
+## [2.2.3] - 2024-04-26
 
 ### Fixed